Update ESP_LEDMatrix_clock_weather.ino
line 402 error

+++ ESP_LEDMatrix_clock_weather/ESP_LEDMatrix_clock_weather.ino
@@ -399,7 +399,7 @@ void getTime()
 void updateTime()
 {
   long curEpoch = localEpoc + ((millis() - localMillisAtUpdate) / 1000);
-  long epoch = round(curEpoch + 3600 * utcOffset + 86400L) % 86400L;
+long epoch = (long)(round(curEpoch + 3600 * utcOffset + 86400L)) % 86400L;
   h = ((epoch  % 86400L) / 3600) % 24;
   m = (epoch % 3600) / 60;
   s = epoch % 60;
